ads/pyproject.toml
[project]
name = "acrique-ads"
version = "0.1.0"
description = "ACRIQUE広告運用CLI - Meta/Google広告をIaCで管理"
requires-python = ">=3.12"
# TODO: 仮置き - 依存関係は実装進行に応じて追加・調整
dependencies = [
    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",           # リッチなターミナル出力

    # 設定管理
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",

    # Meta広告 SDK
    "facebook-business>=19.0.0",

    # Google広告 SDK
    "google-ads>=24.0.0",

    # ユーティリティ
    "httpx>=0.25.0",          # HTTPクライアント
    "python-dotenv>=1.0.0",   # .env読み込み
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# TODO: 仮置き - CLIエントリーポイント
[project.scripts]
ads = "src.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# TODO: 仮置き - Ruff設定
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

# TODO: 仮置き - mypy設定
[tool.mypy]
python_version = "3.12"
strict = true
